[refactor] optimzed folder structure for components
