+++ src/components/candidate/DeleteCandidateModal.jsx
