+++ src/components/election/electionPhaseDetails.jsx
