+++ src/components/nominationDistrict/DeleteDistrictModal.jsx
