+++ src/components/pollingStation/ListPollingStations.jsx
@@ -2,8 +2,8 @@ import { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { getPollingStations } from '../../API';
 import { toast } from 'react-hot-toast';
-import EditPollingStationModal from './EditPollingStationModal';
-import DeleteConfirmationModal from './DeletePollingStationModal';
+import EditPollingStationModal from '../../components/pollingStation/EditPollingStationModal';
+import DeleteConfirmationModal from '../../components/pollingStation/DeletePollingStationModal';
 import { FaEdit, FaTrash } from 'react-icons/fa';
 
 function ListPollingStations({ refresh }) {
